Refactor and modernize Chromecast library tests
Significantly refactored the codebase to improve test structure and enhance functionality. Key updates include:
- Updated test classes to use Xunit Fixtures, V3 and improved output management.
- Implemented nullable reference types for better null handling.
- Standardized async/await usage across the codebase.
- Enhanced event handling for Chromecast discovery with new event args.
- Improved logging with structured messages for better tracking.
- Refined media and queue management functionalities.
- Replaced Tmds.MDns with Zeroconf for mDNS discovery.
- Added XML documentation for clarity on usage.
- Cleaned up unused usings and redundant code.

+++ Sharpcaster/Messages/Queue/QueueInsertMessage.cs
@@ -22,6 +22,6 @@ namespace Sharpcaster.Messages.Queue
         /// </summary>
         [JsonPropertyName("insertBefore")]
         [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-        public long? InsertBefore { get; set; }
+        public int? InsertBefore { get; set; }
     }
 }
